Extract Statistics utility to trino-plugin-toolkit
